requirements.txt
tweepy==4.14.0
praw==7.7.1
facebook-sdk==3.1.0
instaloader==4.10.3
snscrape==0.7.0.20230622
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.2
sqlalchemy==2.0.23
langdetect==1.0.9
textblob==0.17.1
matplotlib==3.8.2
schedule==1.2.0
python-dotenv==1.0.0
nltk==3.8.1
scikit-learn==1.3.2
# TikTokApi==5.3.3  # Commented out - incompatible with Python 3.12
linkedin-api==2.2.1
telethon==1.33.1
discord.py==2.3.2
Mastodon.py==1.8.1
PyGithub==2.1.1
vk-api==11.9.9